+++ CHANGELOG.md
@@ -21,3 +21,9 @@ Initial release to public.
   followed by a PEP 263-conformant character encoding specification
   (see `partial_header.mustache`).  All content currently generated presumes
   UTF-8 encoding, but alternate encodings may be supported in the future.
+
+### Version 0.0.4
+
+* Adds Mustache template for `rest.py` to generate code to use Python 'requests'
+  package instead of the default ('urllib3') as the underlying HTTP client agent,
+  specified by `-D library=requests`.